Add test cases for freestanding `\` followed by escaped reference symbol
Reclass's reference parsing only requires escaping backslashes that
should be literals when they precede a reference opening or closing
symbol. Other backslashes don't need to be escaped. The parser will try
to parse backslashes as single characters first, and will only interpret
them as escape characters when they precede a reference opening or
closing symbol.

We add three test cases which illustrate this behavior: `\\\${foo}`
which is parsed as a single `\` followed by a double-escaped reference,
`\\\\${foo}` which is parsed as two `\` followed by a double-escaped
reference, and `${foo\\\}` which is parsed as a reference to `foo\\`.