Add initial support for parsing Reclass nodes
The commit adds a new struct `Node` which is used to represent Reclass
nodes and classes, code to load YAML files from the filesystem, and a
function which transforms relative class includes into their absolute
form, based on the node's location within the inventory structure.

We also add a very simple example inventory in `tests/inventory` which
will be extended with additional cases as new features are implemented.