Add initial NodeInfo types and expose in Python module
The `NodeInfo` and `NodeInfoMeta` types provide structured forms of the
return value of the Python Reclass `nodeinfo()` method.

We introduce initial versions of `as_py_obj()` and `as_py_dict()` for
the serde_yaml `Value` and `Mapping` types. These won't remain as-is,
but allow us to transform serde_yaml data into Python objects for now.

We also add the `chrono` crate to manage the Node's render timestamp in
Rust. Additionally, we enable the pyo3 `chrono` feature to get
conversion between `chrono::DateTime` and `PyDateTime`.