Add support for extending values in `Mapping::insert()`
We extend our Mapping implementation to support the
`KeyPrefix::Override` marker which allows overriding existing keys
rather than extending them.

Instead of bolting on this support, we change the implementation of
`insert()` to create value lists to keep track of all inserted values by
default, and to only drop the existing value for a key when the new
insertion uses the override key prefix.

While this changes the semantics of `insert()`, the new semantics match
our requirements for Reclass parameter merging better than the previous
semantics. The changed semantics are reflected as changes to some of the
existing unit tests.

See the docstring comments for a detailed explanation of the new
semantics of `Mapping::insert()`.