Track override prefixes which aren't applied during insertion
We need to keep track of override key prefixes which aren't applied
during insertion. The only case where an override isn't applied during
insertion is for keys with the override prefix which don't exist yet in
the mapping.

This case usually occurs when we initially load a Reclass class which
contains an override from disk.