Add support for merging a `Mapping` into another `Mapping`
In order to handle constant keys and overrides in a consistent fashion,
we split out the `insert()` implementation into a private method which
has extra flags which allow the caller to force a key to be constant or
overriding.

We set these flags based on the contents of `const_keys` and
`override_keys` in the Mapping that's getting merged into the target
Mapping to correctly propagate constant keys and unprocessed overrides.

Note that this approach differs from Python Reclass, which tracks
constantness and overrides in the Mapping values and only processes this
information when a `ValueList` is flattened into a single value.

As far as we've been able to determine, our implementation should
produce identical results to Python Reclass, with the possible exception
of some combinations of multiple key prefixes for a single key.