Implement ValueList flattening
This commit introduces `Value::flattened()` which returns a copy of the
Value on which it is called with any ValueLists processed and flattened
into other Value variants. The merging logic for flattening a list of
values is factored into a separate private method `Value::merge()` which
performs the recursive merge of the value list elements.

`Value::merge()` calls `Mapping::merge()` to merge mappings which are
stored in a ValueList. `Mapping::merge()` also calls `flattened()` again
to ensure that any new ValueLists which are created by
`Mapping::merge()` end up flattened in the final result of the original
`flattened()` call.

Finally, the commit also introduces `Value::flatten()` which flattens a
Value in-place, by calling `Value::flattened()` and replacing the
original value with the result of `flattened()`.