Handle symlinks in `nodes_path` or `classes_path` as separate entities
We introduce function `to_lexical_absolute()` which converts a Path into
its normalized absolute form without resolving symlinks. We replace our
usage of `canonicalize()` with that function, so that we correctly treat
symlinks in `nodes_path` or `classes_path` as separate entities (nodes
or classes).

Additionally, we also use `to_lexical_absolute()` when generating the
`uri` in the `NodeInfoMeta` struct in `Node::parse()` to match Python
reclass's behavior.