Implement reference loop detection
We add logic to track any reference paths which we've already seen while
resolving references in a `Value`. Whenever we encounter a reference
path which we've already seen before, we abort reference resolution and
provide the list of paths which form a loop in the error message.

Note that the already introduced boundaries where we don't need to track
paths seen by recursive calls (across ValueList layers, between Sequence
elements and between Mapping key-value pairs) still apply.

We also leave the recursion depth limit in place for now as a safeguard
for potential bugs in the loop detection.

The existing tests are updated to reflect the changed error messages for
the test cases which contain reference loops.