Fix propagation of constant and overriding keys in Mapping interpolation and flattening
We introduce a new internal method `Mapping::flattened()` which is
called from `Value::flattened()` for Mapping values. This method ensures
that the flattened mapping retains the constant and overriding key
information from the original mapping.

We also ensure that constant and overriding key information is
propagated to the new mapping in `Mapping::interpolate()`.

Additionally, the commit updates all the node rendering test cases to
have the correct sets of constant and overriding keys in the expected
final parameters data.