Improve error messages when reference resolution fails
To improve the error messages for reference resolution failure, we add a
`Vec<String>` to track the current parameter key for which we're
resolving references into the `ResolveState` struct.

We also extract the error rendering in `Token::resolve()` into methods
on `ResolveState`, since we mostly need data from `ResolveState` to
render the error messages.

The new error messages all print the parameter key for which we're
resolving references. If applicable, the error messages also contain the
full reference, and the particular key for which resolution failed.

To reduce repetition for the error handling, we also refactor
`Token::resolve()` to always either interpolate or copy the Value which
is the base for the key lookup. While this introduces a bunch of
unnecessary calls to `clone()` if we're not doing a lookup into a
`Value::String` or `Value::ValueList`, the performance impact is
negligible, and the resulting code is much less convoluted.

Resolves #69