Implement path options which match Python reclass
This commit introduces a new config option `inventory_path`.
The semantics of the `nodes_path` and `classes_path` options are
adapted to match Python reclass's behavior, so that reclass-rs now
expects `nodes_path` and `classes_path` to be specified as relative
paths to `inventory_path`.

We default `inventory_path` to the current directory, the same as Python
Reclass in some cases, so users can provide concatenated paths for
`nodes_path` and `classes_path`.  We choose this default, so that we can
use Kapitan's prerendered values for `nodes_path` and `classes_path`.