Move logic to compute class "location" to initial class discovery
This improves performance for inventories which include the same classes
in many places, since we don't have to redo the same work to determine
the class's location every time it's included.

The class location is used to turn relative into absolute imports.

We introduce a new struct `EntityInfo` which we use to track an entity's
path and "location".