Treat `foo/init.yml` as content for class `foo`
This ensures that the behavior of reclass-rs matches Python reclass in
regards to how files named `init.yml` (or `init.yaml`) are treated.

Python reclass treats such files as content for class `foo`. While
Project Syn doesn't make use of this mapping and instead favors having a
file `foo.yml` alongside a folder `foo/` to achieve the same behavior,
we want to ensure that reclass-rs is compatible with as many Reclass
inventories as possible.

Notably, Python reclass also resolves relative includes in such files as
relative to the parent directory of the directory containing the file.
For example, relative includes in `foo/init.yml` are treated relative to
the directory holding `foo` and not relative to directory `foo`.

This is a breaking change for users which rely on the current reclass-rs
behavior of handling files named `init.yml`, since such files can no
longer be included as `foo.init`.