Implement Reclass option compose-node-name
When enabled, this option allows users to define nodes in nested
directories in `nodes_path`. Node names are constructed the same way as
class names, by replacing directory separators with dots.

This commit implements support for compose-node-name in such a way that
the resulting node metadata is compatible with Python reclass.