Implement compat flag for node metadata when compose-node-name is enabled
Python reclass discards literal dots in node filepaths when
compose-node-name is enabled, resulting in incorrect node metadata for
nodes which have literal dots in their filename (e.g. node `a.1` in the
`test/inventory-compose-node-name` inventory).

This commit implements compatibility flag `ComposeNodeNameLiteralDots`.
When set, this flag tells reclass-rs to generate node metadata which
matches Python reclass when compose-node-name is enabled.

By default (i.e. without the compatibility flag), reclass-rs preserves
literal dots in node names when compose-node-name is enabled.

To implement the new default behavior, we additionally track the node's
path in `nodes_path` in `NodeInfoMeta`, so that we can ensure that
literal dots are preserved when generating the node's metadata.