Make `ignore_class_notfound_regexp` config fields private
Instead of exposing the fields (which must be kept in sync) directly, we
expose a getter & setter pair. The setter ensures that a new RegexSet is
compiled after the internal `ignore_class_notfound_regexp` field is
updated.

We still expose the field read-only to Python through a generated PyO3
getter.