Add Python tests for class mappings
We write tests for the `class_mappings` inventory by simply running
reclass-rs and Python reclass with the same config and verifying that
the output matches.

To ensure that we don't run into issues, we delete all timestamps from
the reclass-rs and kapicorp-reclass output. Additionally, we sort the
values for the top-level `classes` and `applications` dictionaries for
the kapicorp-reclass output, since those aren't necessarily sorted by
the Python implementation.

Finally, skip the new tests on Windows, since kapicorp-reclass doesn't
support Windows. To ensure that the tests can be skipped, we only import
`reclass` and `reclass.core` inside the test function.