Extend test inventory to validate shown class order
We originally had a slightly different approach to include classes that
are included via `class_mappings` which caused nested includes to appear
in a different order in the inventory output when comparing between
Python reclass and reclass-rs.

The latest version of the implementation is adjusted to use the exact
same approach as Python reclass when `class_mappings` are present.

This commit extends the `class_mappings` test inventory to cover this
case in the direct comparison Python test.