Add `buildinfo()` Python method
This currently returns a dict with a single entry for the rustc version.

We do this because the Rust compiler doesn't embed its own version into
the resulting library on macOS (not sure about Windows), so we can't
reliably extract the compiler version that was used to build reclass-rs
from the library `.so`.

The implementation uses a Rust build script (`build.rs`) which uses the
`rustc_version` crate to get the Rust compiler's version. For now, the
crate just provides a convenience wrapper for calling `rustc --version`,
but would allow us to access more structured compiler info in the
future.