Update release action to not use deprecated `::set-output`
We also need to adjust how we pass the release summary that we read from
the Git tag to `ncipollo/release-action`: when passing step outputs via
`$GITHUB_OUTPUT` we no longer need to escape newlines with `%0A`, but
instead we need to encode the whole message as JSON to preserve
newlines.

We then apply the GitHub actions builtin `fromJSON()` to decode the
message when passing it to `ncipollo/release-action` for the release
body.

Note that the updated `sed` expression now directly operates on
newlines, and also preserves list items as long as the list item's '*'
is directly at the start of the line.