Implement `ignore_overwritten_missing_references` reclass config option
We default the option to `true` to retain the kapicorp-reclass default
behavior. In contrast to kapicorp-reclass, we can't currently collect
all resolution errors but instead hard-fail on the first non-ignored
error we encounter.

Technically this is a breaking change for existing reclass-rs
inventories.

The implementation introduces a new Value variant `ResolveError` which
is used by the Value implementation to decide which resolve errors are
always errors and which can be downgraded to warnings when Reclass
option `ignore_overwritten_missing_references` is set to `true`.

Fixes #184