Use if-let chains where it makes sense
Rust 2024 edition has stabilized [if-let chains]. We start using if-let
chains for code blocks where we previously had to use suboptimal
structure (e.g. in `coalesce_literals()` in the ref parser) or where we
can avoid using `unwrap()` in if blocks which make the `unwrap()`
infallible (e.g. some `ignore_overwritten_missing_references` handling).

Some of the changes in this commit also make use of the [if-let scope]
change in the 2024 edition, which previously prevented using `if let`
expressions in some cases where we actually want to drop the borrowed
value for the `else`.

[if-let chains]: https://doc.rust-lang.org/edition-guide/rust-2024/let-chains.html
[if-let scope]: https://doc.rust-lang.org/edition-guide/rust-2024/temporary-if-let-scope.html