Update pyproject.toml `dynamic` entries to ensure full Python package metadata
We extend the `dynamic` field in `pyproject.toml` to ensure that the
Python package is built with all the metadata defined in `Cargo.toml`.

The behavior of `maturin` has changed to omit metadata defined in
`Cargo.toml` unless the corresponding `pyproject.toml` fields are
specifed as `dynamic`, cf. https://github.com/PyO3/maturin/pull/2672